Optimized my job template
The biggest change is using MY.Arguments instead of Arguments, which
is a classad expression instead of a string. This let me get away with
omitting all kinds of quoting.

At this point, using condor_qedit on any of the Dask* attributes
_should_ cause the corresponding changes to the arguments list (though
I haven't tested that yet).

Thanks to TJ for lots of help.